requirements.txt
requests==2.32.3
pillow==10.4.0
lxml==5.2.2
qrcode==8.0
psutil==6.1.1
APScheduler==3.10.4
urllib3==2.2.2
pixie-python==4.3.0
qq-botpy==1.2.1
numpy==2.2.1
pypinyin==0.53.0
aiohttp==3.11.0
easyocr==1.7.2
nest-asyncio==1.6.0
thefuzz==0.22.1
easy-pixie==0.0.6
markdown2==2.5.3
imgkit==1.2.3